feat: refactor large components into smaller units for better maintainability
